+++ src/deluge/storage/flash_storage.cpp
@@ -179,6 +179,7 @@ enum Entries {
 171: default swing interval
 172: default disabled scales low byte
 173: default disabled scales high byte
+174: default performance view gold knob value editng
 */
 
 uint8_t defaultScale;
@@ -215,6 +216,8 @@ bool automationClear = true;
 bool automationShift = true;
 bool automationNudgeNote = true;
 bool automationDisableAuditionPadShortcuts = true;
+
+bool performanceGoldKnobValueEditing = false;
 
 StartupSongMode defaultStartupSongMode;
 
@@ -326,6 +329,8 @@ void resetSettings() {
 	defaultSwingInterval = 8 - defaultMagnitude; // 16th notes
 
 	defaultDisabledPresetScales = {0};
+
+	performanceGoldKnobValueEditing = false;
 }
 
 void resetMidiFollowSettings() {
@@ -703,6 +708,13 @@ void readSettings() {
 	else {
 		defaultDisabledPresetScales = std::bitset<NUM_PRESET_SCALES>((buffer[173] << 8) | buffer[172]);
 	}
+
+	if (buffer[174] != 0 && buffer[174] != 1) {
+		performanceGoldKnobValueEditing = false;
+	}
+	else {
+		performanceGoldKnobValueEditing = buffer[174];
+	}
 }
 
 static bool areMidiFollowSettingsValid(std::span<uint8_t> buffer) {
@@ -965,6 +977,8 @@ void writeSettings() {
 	unsigned long disabledBits = defaultDisabledPresetScales.to_ulong();
 	buffer[172] = 0xff & disabledBits;
 	buffer[173] = 0xff & (disabledBits >> 8);
+
+	buffer[174] = performanceGoldKnobValueEditing;
 
 	R_SFLASH_EraseSector(0x80000 - 0x1000, SPIBSC_CH, SPIBSC_CMNCR_BSZ_SINGLE, 1, SPIBSC_OUTPUT_ADDR_24);
 	R_SFLASH_ByteProgram(0x80000 - 0x1000, buffer.data(), 256, SPIBSC_CH, SPIBSC_CMNCR_BSZ_SINGLE, SPIBSC_1BIT,